+++ src/main/java/com/fortanix/sdkms/performance/sampler/AbstractMACSampler.java
@@ -11,10 +11,7 @@ package com.fortanix.sdkms.performance.sampler;
 import com.fortanix.sdkms.v1.ApiException;
 import com.fortanix.sdkms.v1.api.DigestApi;
 import com.fortanix.sdkms.v1.api.SecurityObjectsApi;
-import com.fortanix.sdkms.v1.model.DigestAlgorithm;
-import com.fortanix.sdkms.v1.model.MacGenerateRequest;
-import com.fortanix.sdkms.v1.model.ObjectType;
-import com.fortanix.sdkms.v1.model.SobjectRequest;
+import com.fortanix.sdkms.v1.model.*;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.jmeter.protocol.java.sampler.JavaSamplerContext;
 
@@ -22,6 +19,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.security.ProviderException;
+import java.util.List;
 import java.util.UUID;
 import java.util.logging.Level;
 
@@ -42,6 +40,7 @@ public abstract class AbstractMACSampler extends AbstractSDKMSSamplerClient {
         String keySize = context.getParameter(KEY_SIZE, "160");
         String hashAlgorithm = context.getParameter(HASH_ALGORITHM, "SHA1");
         String filePath = context.getParameter(FILE_PATH);
+        String keyName = context.getParameter(KEYNAME, "");
         ObjectType objectType = ObjectType.fromValue(algorithm);
         this.hashAlgorithm = DigestAlgorithm.fromValue(hashAlgorithm);
         String input = "random-text";
@@ -56,12 +55,28 @@ public abstract class AbstractMACSampler extends AbstractSDKMSSamplerClient {
         this.text = input;
         super.setupTest(context);
         this.securityObjectsApi = new SecurityObjectsApi(this.apiClient);
-        SobjectRequest sobjectRequest = new SobjectRequest().name(UUID.randomUUID().toString()).objType(objectType).keySize(Integer.parseInt(keySize));
-        try {
-            this.keyId = this.securityObjectsApi.generateSecurityObject(sobjectRequest).getKid();
-        } catch (ApiException e) {
-            LOGGER.log(Level.INFO, "failure in creating key : " + e.getMessage(), e);
-            throw new ProviderException(e.getMessage());
+        List<KeyObject> sObjects;
+        KeyObject matched = null;
+        if(keyName != ""){
+            try {
+                sObjects = securityObjectsApi.getSecurityObjects(keyName, null, null, null, true, null, null, null);
+                matched = sObjects.get(0);
+            }
+            catch(ApiException e){
+                LOGGER.log(Level.INFO, "failure in getting keyObject from keyName: " + e.getMessage(), e);
+            }
+        }
+        if(matched != null){
+            this.keyId = matched.getKid();
+        }
+        else {
+            SobjectRequest sobjectRequest = new SobjectRequest().name(UUID.randomUUID().toString()).objType(objectType).keySize(Integer.parseInt(keySize));
+            try {
+                this.keyId = this.securityObjectsApi.generateSecurityObject(sobjectRequest).getKid();
+            } catch (ApiException e) {
+                LOGGER.log(Level.INFO, "failure in creating key : " + e.getMessage(), e);
+                throw new ProviderException(e.getMessage());
+            }
         }
         this.macGenerateRequest = new MacGenerateRequest().alg(this.hashAlgorithm).data(this.text.getBytes());
         this.digestApi = new DigestApi(this.apiClient);
@@ -69,10 +84,13 @@ public abstract class AbstractMACSampler extends AbstractSDKMSSamplerClient {
 
     @Override
     public void teardownTest(JavaSamplerContext context) {
-        try {
-            this.securityObjectsApi.deleteSecurityObject(this.keyId);
-        } catch (ApiException e) {
-            LOGGER.log(Level.INFO, "failure in deleting key : " + e.getMessage(), e);
+        String keyName = context.getParameter(KEYNAME, "");
+        if(keyName == ""){
+            try {
+                this.securityObjectsApi.deleteSecurityObject(this.keyId);
+            } catch (ApiException e) {
+                LOGGER.log(Level.INFO, "failure in deleting key : " + e.getMessage(), e);
+            }
         }
         super.teardownTest(context);
     }
